[dev] reformatted the readme
This was originally paragraph-filled in vi, and I'm redoing it in
Emacs (which uses a different column).

I removed the section about how I couldn't find a way to preserve the
Git history of a subdirectory of the original project.  I've since
learned that is an easy thing to do.  Anyway, it's so long ago that
there's no point in mentioning now.